Chore: Move CSS file to new location
