+++ src/pages/books/BooksHomePage.tsx
@@ -1,8 +1,8 @@
+import styles from './styles/BooksHomePage.module.css';
 import BestSellerSection from "../../features/books/components/BookBestSeller";
 import BookSlider from "../../features/books/components/BookSlider";
 import BookSectionHeader from "../../features/books/components/BookSectionHeader";
 import CategoryFilterSearchBar from "../../features/books/components/CategoryFilterSearchBar";
-import styles from './BooksHomePage.module.css';
 import BookCardForBookSlider from "../../features/books/components/BookCardForBookSlider";
 import Banner from "../../components/display/BannerSlider";
 import Footer from "../../components/layout/Footer";
